+++ packages/aws-amplify-react/src/Amplify-UI/Amplify-UI-Theme.jsx
@@ -11,34 +11,40 @@
  * and limitations under the License.
  */
 
-export const Container = {}
-export const FormContainer = {}
-export const FormSection = {}
-export const FormField = {}
-export const SectionHeader = {}
-export const SectionBody = {}
-export const SectionFooter = {}
-export const SectionFooterPrimaryContent = {}
-export const SectionFooterSecondaryContent = {}
-export const Input = {}
-export const Button = {}
+export const Container = {};
+export const FormContainer = {};
+export const FormSection = {};
+export const FormField = {};
+export const SectionHeader = {};
+export const SectionBody = {};
+export const SectionFooter = {};
+export const SectionFooterPrimaryContent = {};
+export const SectionFooterSecondaryContent = {};
+export const Input = {};
+export const Button = {};
 export const PhotoPickerButton = {};
 export const PhotoPlaceholder = {};
-export const SignInButton = {}
-export const SignInButtonIcon = {}
-export const SignInButtonContent = {}
-export const Strike = {}
-export const StrikeContent = {}
-export const ActionRow = {}
-export const FormRow = {}
-export const A = {}
-export const Hint = {}
-export const Radio = {}
-export const InputLabel = {}
+export const SignInButton = {};
+export const SignInButtonIcon = {};
+export const SignInButtonContent = {};
+export const Strike = {};
+export const StrikeContent = {};
+export const ActionRow = {};
+export const FormRow = {};
+export const A = {};
+export const Hint = {};
+export const Radio = {};
+export const InputLabel = {};
 export const AmazonSignInButton = {};
 export const FacebookSignInButton = {};
 export const GoogleSignInButton = {};
 export const OAuthSignInButton = {};
+export const Toast = {};
+export const NavBar = {};
+export const NavRight = {};
+export const Nav = {};
+export const NavItem = {};
+export const NavButton = {};
 
 const AmplifyTheme = {
   container: Container,
@@ -72,7 +78,14 @@ const AmplifyTheme = {
 
   hint: Hint,
   radio: Radio,
-  inputLabel: InputLabel
-}
+  inputLabel: InputLabel,
+  toast: Toast,
+  
+  navBar: NavBar,
+  nav: Nav,
+  navRight: NavRight,
+  navItem: NavItem,
+  navButton: NavButton
+};
 
 export default AmplifyTheme;