+++ packages/aws-amplify-react/src/Auth/FederatedSignIn.jsx
@@ -15,7 +15,8 @@ import {
     GoogleButton,
     FacebookButton,
     AmazonButton,
-    OAuthButton
+    OAuthButton,
+    Auth0Button
 } from './Provider';
 
 const logger = new Logger('FederatedSignIn');
@@ -29,7 +30,7 @@ export class FederatedButtons extends Component {
                 google_client_id={google_client_id}
                 theme={theme}
                 onStateChange={onStateChange}
-              />
+              />;
     }
 
     facebook(facebook_app_id) {
@@ -40,7 +41,7 @@ export class FederatedButtons extends Component {
                 facebook_app_id={facebook_app_id}
                 theme={theme}
                 onStateChange={onStateChange}
-                />
+                />;
     }
 
     amazon(amazon_client_id) {
@@ -51,7 +52,7 @@ export class FederatedButtons extends Component {
                 amazon_client_id={amazon_client_id}
                 theme={theme}
                 onStateChange={onStateChange}
-              />
+              />;
     }
 
     OAuth(oauth_config) {
@@ -61,6 +62,17 @@ export class FederatedButtons extends Component {
                 label={oauth_config? oauth_config.label : undefined}
                 theme={theme}
                 onStateChange={onStateChange}
+              />;
+    }
+
+    auth0(auth0) {
+        if (!auth0) { return null;}
+        const { theme, onStateChange } = this.props;
+        return <Auth0Button
+                label={auth0? auth0.label : undefined}
+                theme={theme}
+                onStateChange={onStateChange}
+                auth0={auth0}
               />
     }
 
@@ -73,14 +85,21 @@ export class FederatedButtons extends Component {
             throw new Error('No Auth module found, please ensure @aws-amplify/auth is imported');
         }
 
-        const config = Auth.configure();
-        if (config.oauth) {
-            federated.oauth_config = Object.assign({}, federated.oauth_config, config.oauth);
+        const { oauth={} } = Auth.configure();
+        // backward compatibility
+        if (oauth['domain']) {
+            federated.oauth_config = Object.assign({}, federated.oauth_config, oauth);
+        } else if (oauth.awsCognito) {
+            federated.oauth_config = Object.assign({}, federated.oauth_config, oauth.awsCognito);
+        }
+
+        if (oauth.auth0) {
+            federated.auth0 = Object.assign({}, federated.auth0, oauth.auth0);
         }
 
         if (JS.isEmpty(federated)) { return null; }
 
-        const { google_client_id, facebook_app_id, amazon_client_id, oauth_config } = federated;
+        const { google_client_id, facebook_app_id, amazon_client_id, oauth_config, auth0 } = federated;
 
         const theme = this.props.theme || AmplifyTheme;
         return (
@@ -97,23 +116,33 @@ export class FederatedButtons extends Component {
                 <div>
                 {this.OAuth(oauth_config)}
                 </div>
-                <Strike>{I18n.get('or')}</Strike>
+                <div>
+                {this.auth0(auth0)}
+                </div>
+                <Strike theme={theme}>{I18n.get('or')}</Strike>
             </div>
-        )
+        );
     }
 }
 
 export default class FederatedSignIn extends Component {
     render() {
         const { authState, onStateChange } = this.props;
-        let federated = this.props.federated || {};
+        const federated = this.props.federated || {};
         if (!Auth || typeof Auth.configure !== 'function') {
             throw new Error('No Auth module found, please ensure @aws-amplify/auth is imported');
         }
         
-        const config = Auth.configure();
-        if (config.oauth) {
-            federated.oauth_config = Object.assign({}, federated.oauth_config, config.oauth);
+         const { oauth={} } = Auth.configure();
+        // backward compatibility
+        if (oauth['domain']) {
+            federated.oauth_config = Object.assign({}, federated.oauth_config, oauth);
+        } else if (oauth.awsCognito) {
+            federated.oauth_config = Object.assign({}, federated.oauth_config, oauth.awsCognito);
+        }
+
+        if (oauth.auth0) {
+            federated.auth0 = Object.assign({}, federated.auth0, oauth.auth0);
         }
 
         if (!federated) {
@@ -135,6 +164,6 @@ export default class FederatedSignIn extends Component {
                     />
                 </SectionBody>
             </FormSection>
-        )
+        );
     }
 }