+++ packages/aws-amplify-react/src/Auth/Provider/withGoogle.jsx
@@ -1,3 +1,16 @@
+/*
+ * Copyright 2017-2018 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"). You may not use this file except in compliance with
+ * the License. A copy of the License is located at
+ *
+ *     http://aws.amazon.com/apache2.0/
+ *
+ * or in the "license" file accompanying this file. This file is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR
+ * CONDITIONS OF ANY KIND, either express or implied. See the License for the specific language governing permissions
+ * and limitations under the License.
+ */
+
 import * as React from 'react';
 import { Component } from 'react';
 
@@ -10,7 +23,7 @@ import {
     SignInButtonIcon,
     SignInButtonContent
 } from '../../Amplify-UI/Amplify-UI-Components-React';
-
+import Constants from '../common/constants';
 
 const logger = new Logger('withGoogle');
 
@@ -21,6 +34,7 @@ export default function withGoogle(Comp) {
 
             this.initGapi = this.initGapi.bind(this);
             this.signIn = this.signIn.bind(this);
+            this.signOut = this.signOut.bind(this);
             this.federatedSignIn = this.federatedSignIn.bind(this);
 
             this.state = {};
@@ -30,7 +44,18 @@ export default function withGoogle(Comp) {
             const ga = window.gapi.auth2.getAuthInstance();
             const { onError } = this.props;
             ga.signIn().then(
-                googleUser => this.federatedSignIn(googleUser),
+                googleUser => {
+                    this.federatedSignIn(googleUser);
+                    const payload = {
+                        provider: Constants.GOOGLE
+                    }
+
+                    try {
+                        localStorage.setItem(Constants.AUTH_SOURCE_KEY, JSON.stringify(payload));
+                    } catch (e) {
+                        logger.debug('Failed to cache auth source into localStorage', e);
+                    }
+                },
                 error => {
                     if (onError) onError(error);
                     else throw error;
@@ -66,9 +91,26 @@ export default function withGoogle(Comp) {
             }
         }
 
+        signOut() {
+            const authInstance = window.gapi && window.gapi.auth2? window.gapi.auth2.getAuthInstance() : null;
+            if (!authInstance) {
+                return Promise.resolve();
+            }
+
+            authInstance.then((googleAuth) => {
+                if (!googleAuth) {
+                    logger.debug('google Auth undefined');
+                    return Promise.resolve();
+                }
+
+                logger.debug('google signing out');
+                return googleAuth.signOut();
+            });
+        }
+
         componentDidMount() {
             const { google_client_id } = this.props;
-            if (google_client_id) this.createScript();
+            if (google_client_id && !window.gapi) this.createScript();
         }
 
         createScript() {
@@ -98,7 +140,7 @@ export default function withGoogle(Comp) {
                 window.gapi && window.gapi.auth2
                     ? window.gapi.auth2.getAuthInstance()
                     : null;
-            return <Comp {...this.props} ga={ga} googleSignIn={this.signIn} />;
+            return <Comp {...this.props} ga={ga} googleSignIn={this.signIn} googleSignOut={this.signOut} />;
         }
     };
 }