+++ packages/aws-amplify-vue/src/components/index.js
@@ -1,3 +1,4 @@
+export * from './api';
 export * from './authenticator';
 export * from './storage';
 export * from './interactions';