+++ packages/aws-amplify-vue/CHANGELOG.md
@@ -3,6 +3,73 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.1.7-unstable.1"></a>
+## [0.1.7-unstable.1](https://github.com/aws-amplify/amplify-js/compare/aws-amplify-vue@0.1.7-unstable.0...aws-amplify-vue@0.1.7-unstable.1) (2018-11-09)
+
+
+
+
+**Note:** Version bump only for package aws-amplify-vue
+
+<a name="0.1.7-unstable.0"></a>
+## [0.1.7-unstable.0](https://github.com/aws-amplify/amplify-js/compare/aws-amplify-vue@0.1.6...aws-amplify-vue@0.1.7-unstable.0) (2018-11-07)
+
+
+
+
+**Note:** Version bump only for package aws-amplify-vue
+
+<a name="0.1.6"></a>
+## [0.1.6](https://github.com/aws-amplify/amplify-js/compare/aws-amplify-vue@0.1.6-unstable.0...aws-amplify-vue@0.1.6) (2018-11-01)
+
+
+
+
+**Note:** Version bump only for package aws-amplify-vue
+
+<a name="0.1.6-unstable.0"></a>
+## [0.1.6-unstable.0](https://github.com/aws-amplify/amplify-js/compare/aws-amplify-vue@0.1.5...aws-amplify-vue@0.1.6-unstable.0) (2018-10-30)
+
+
+
+
+**Note:** Version bump only for package aws-amplify-vue
+
+<a name="0.1.5"></a>
+## [0.1.5](https://github.com/aws-amplify/amplify-js/compare/aws-amplify-vue@0.1.5-unstable.1...aws-amplify-vue@0.1.5) (2018-10-29)
+
+
+
+
+**Note:** Version bump only for package aws-amplify-vue
+
+<a name="0.1.5-unstable.1"></a>
+## [0.1.5-unstable.1](https://github.com/aws-amplify/amplify-js/compare/aws-amplify-vue@0.1.5-unstable.0...aws-amplify-vue@0.1.5-unstable.1) (2018-10-29)
+
+
+
+
+**Note:** Version bump only for package aws-amplify-vue
+
+<a name="0.1.5-unstable.0"></a>
+## [0.1.5-unstable.0](https://github.com/aws-amplify/amplify-js/compare/aws-amplify-vue@0.1.4...aws-amplify-vue@0.1.5-unstable.0) (2018-10-24)
+
+
+### Features
+
+* **aws-amplify-vue:** add Connect component ([ae3ac0a](https://github.com/aws-amplify/amplify-js/commit/ae3ac0a))
+
+
+
+
+<a name="0.1.4"></a>
+## [0.1.4](https://github.com/aws-amplify/amplify-js/compare/aws-amplify-vue@0.1.4-unstable.1...aws-amplify-vue@0.1.4) (2018-10-17)
+
+
+
+
+**Note:** Version bump only for package aws-amplify-vue
+
 <a name="0.1.4-unstable.1"></a>
 ## [0.1.4-unstable.1](https://github.com/aws-amplify/amplify-js/compare/aws-amplify-vue@0.1.4-unstable.0...aws-amplify-vue@0.1.4-unstable.1) (2018-10-10)